fix(plasma-ui): circular dependency
