+++ packages/ui/src/components/Toast/useToast.ts
@@ -1,6 +1,6 @@
 import { useContext } from 'react';
 
-import { ToastContext } from './ToastProvider';
+import { ToastContext } from './ToastContext';
 
 export const useToast = () => {
     const { showToast, hideToast } = useContext(ToastContext);
